+++ exampleCourse/questions/brokenGrading/server.py
@@ -1,0 +1,7 @@
+import random, copy
+
+def generate(data):
+    data['correct_answers']['x'] = 3
+
+def grade(data):
+    raise Exception('deliberately broken grading function')